requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
chardet==3.0.3
discord.py==0.16.8
lorem-ipsum-generator==0.3
multidict==2.1.5
six==1.10.0
transliterate==1.9
urllib3==1.21.1
websockets==3.3
